test: refactor `mockPrismicRestAPIV2()` into one file
