+++ test/__testutils__/mockPrismicRestAPIV2.ts
@@ -1,73 +1,79 @@
+import { rest } from "msw";
 import { expect, TestContext } from "vitest";
-import {
-	createMSWQueryHandler,
-	CreateMSWQueryHandlerArgs,
-} from "./createMSWQueryHandler";
-import {
-	createMSWRepositoryHandler,
-	CreateMSWRepositoryHandlerArgs,
-} from "./createMSWRepositoryHandler";
+import * as prismicT from "@prismicio/types";
 
 import { createRepositoryName } from "./createRepositoryName";
 
 type MockPrismicRestAPIV2Args = {
 	ctx: TestContext;
-	repositoryName?: string;
 	accessToken?: string;
-	apiEndpoint?: string;
-	repositoryResponse?: CreateMSWRepositoryHandlerArgs["response"];
-	repositoryDuration?: number;
-	queryResponse?: CreateMSWQueryHandlerArgs["response"];
+	repositoryResponse?: prismicT.Repository;
+	queryResponse?: prismicT.Query | prismicT.Query[];
 	queryRequiredParams?: Record<string, string | string[]>;
 	queryDelay?: number;
 };
 
 export const mockPrismicRestAPIV2 = (args: MockPrismicRestAPIV2Args) => {
-	const repositoryName = args.repositoryName || createRepositoryName();
+	const repositoryName = createRepositoryName();
+	const repositoryEndpoint = `https://${repositoryName}.cdn.prismic.io/api/v2`;
+	const queryEndpoint = new URL(
+		"documents/search",
+		repositoryEndpoint + "/",
+	).toString();
 
 	args.ctx.server.use(
-		createMSWRepositoryHandler({
-			ctx: args.ctx,
-			repositoryName,
-			apiEndpoint: args.apiEndpoint,
-			response: (req, res, ctx) => {
-				if (
-					typeof args.accessToken === "string" &&
-					req.url.searchParams.get("access_token") !== args.accessToken
-				) {
-					return res(
-						ctx.status(401),
-						ctx.json({
-							message: "invalid access token",
-							oauth_initiate: "oauth_initiate",
-							oauth_token: "oauth_token",
-						}),
+		rest.get(repositoryEndpoint, (req, res, ctx) => {
+			if (
+				typeof args.accessToken === "string" &&
+				req.url.searchParams.get("access_token") !== args.accessToken
+			) {
+				return res(
+					ctx.status(401),
+					ctx.json({
+						message: "invalid access token",
+						oauth_initiate: "oauth_initiate",
+						oauth_token: "oauth_token",
+					}),
+				);
+			}
+
+			const response =
+				args.repositoryResponse || args.ctx.mock.api.repository();
+
+			return res(ctx.json(response));
+		}),
+		rest.get(queryEndpoint, (req, res, ctx) => {
+			if (args.queryRequiredParams) {
+				for (const paramKey in args.queryRequiredParams) {
+					const requiredValue = args.queryRequiredParams[paramKey];
+
+					expect(req.url.searchParams.getAll(paramKey)).toStrictEqual(
+						Array.isArray(requiredValue) ? requiredValue : [requiredValue],
+					);
+				}
+			}
+
+			const response = args.queryResponse || args.ctx.mock.api.query();
+
+			if (Array.isArray(response)) {
+				const page = Number.parseInt(req.url.searchParams.get("page") || "1");
+
+				const thisResponse = response[page - 1];
+
+				if (!thisResponse) {
+					throw new Error(
+						`A query response was not generated for \`page=${page}\`.`,
 					);
 				}
 
-				return typeof args.repositoryResponse === "function"
-					? args.repositoryResponse(req, res, ctx)
-					: args.repositoryResponse;
-			},
-			delay: args.repositoryDuration,
-		}),
-		createMSWQueryHandler({
-			ctx: args.ctx,
-			repositoryName,
-			apiEndpoint: args.apiEndpoint,
-			validator: (req) => {
-				if (args.queryRequiredParams) {
-					for (const paramKey in args.queryRequiredParams) {
-						const requiredValue = args.queryRequiredParams[paramKey];
-
-						expect(req.url.searchParams.getAll(paramKey)).toStrictEqual(
-							Array.isArray(requiredValue) ? requiredValue : [requiredValue],
-						);
-					}
+				return res(ctx.delay(args.queryDelay), ctx.json(thisResponse));
+			} else {
+				if ("status" in response) {
+					return response;
+				} else {
+					return res(ctx.delay(args.queryDelay), ctx.json(response));
 				}
-			},
-			response: args.queryResponse,
-			delay: args.queryDelay,
+			}
 		}),
 	);
 };
